Fix resetState() function and update xpi
resetState() was only caring about the changes in the hidden array. Make it try
it's hardest to first take care of the changes in the hidden array and only
*then* try to reset the tiles array. It's because hidden array changes are more
important than changes in the tiles array. Update XPI because this fixes a
pretty big bug so that the new users would not hit this bug